- mount memory routes when enabled - pipeline feature flag
