+++ model/app.py
@@ -1,14 +1,6 @@
 from args import args
-from flask import Flask, request
-from flask_cors import CORS
-
-
-# from args import args
-
-app = Flask("agnai-pipeline")
-app.run(host="localhost", port=5001)
-
-CORS(app)
+from flask import request
+from server import app
 
 
 @app.get("/status")
@@ -38,37 +30,11 @@ def summarizePost():
     return {"success": True, "summary": resp}
 
 
-@app.post("/memory/<chat_id>/embed")
-def memoryEmbed(chat_id):
-    payload = request.get_json(Silent=True)
-    messages = payload.get("messages", False) or None
-
-    if messages is None:
-        return {"success": False, "error": "chatId or messages missing in payload"}
-
-    documents = [msg["msg"] for msg in messages]
-    ids = [msg["_id"] for msg in messages]
-    metadatas = [{"name": msg["name"], "date": msg["createdAt"]} for msg in messages]
-
-    collection = memory.client.get_or_create_collection(
-        name=chat_id, embedding_function=memory.embed
-    )
-    collection.upsert(ids=ids, documents=documents, metadatas=metadatas)
-
-    return {"success": True}
-
-
-@app.post("/memory/<chat_id>")
-def memoryRecall(chat_id):
-    payload = request.get_json(Silent=True)
-    collection = memory.client.get_or_create_collection(
-        name=chat_id, embedding_function=memory.embed
-    )
+import memory
+import summary
 
 
 print(
     'Pipeline API started: Remember to enable "USE LOCAL PIPELINE" in your Agnaistic account settings'
 )
-
-import memory
-import summary
+app.run(host="localhost", port=5001)
