+++ model/server.py
@@ -1,0 +1,7 @@
+from flask import Flask, request
+from flask_cors import CORS
+
+app = Flask("agnai-pipeline")
+
+
+CORS(app)
