+++ web/pages/Settings/AISettings.tsx
@@ -67,12 +67,14 @@ const AISettings: Component<{
           value={state.user?.defaultPreset || ''}
         />
 
-        <Toggle
-          fieldName="useLocalPipeline"
-          label="Use Local Pipeline"
-          helperText="If available, use local Agnaistic pipeline features (summarization for images). This is extremely new and experimental. Expect this to change and improve in the near future."
-          value={state.user?.useLocalPipeline}
-        />
+        <Show when={cfg.flags.pipeline}>
+          <Toggle
+            fieldName="useLocalPipeline"
+            label="Use Local Pipeline"
+            helperText="If available, use local Agnaistic pipeline features (summarization for images). This is extremely new and experimental. Expect this to change and improve in the near future."
+            value={state.user?.useLocalPipeline}
+          />
+        </Show>
 
         <div class="my-2">
           <Tabs tabs={tabs()} selected={tab} select={setTab} />
